dac/requirements.txt
django
psycopg2
mysqlclient
djangorestframework
markdown
django-filter
requests
